requirements.txt
asgiref==3.3.4
attrs==21.2.0
boto3==1.17.69
botocore==1.20.69
coverage==5.5
dj-database-url==0.5.0
Django==3.2.2
django-ckeditor==6.0.0
django-environ==0.4.5
django-extensions==3.1.3
django-js-asset==1.2.2
django-storages==1.11.1
iniconfig==1.1.1
jmespath==0.10.0
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4